requirements.txt
asgiref==3.8.0
Django==4.2
django-cors-headers==3.14.0
django-jazzmin==2.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
PyJWT==2.6.0
pytz==2024.1
PyYAML==6.0.1
shortuuid==1.0.13
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
